broke up on test file so we won't get timeouts on travis
